refactor(CB-264): diary 페이지 폴더 구조 변경
